Proptest count

+++ rust/segment/src/sqlite_metadata.rs
@@ -67,7 +67,7 @@ impl SqliteMetadataWriter {
                     .do_nothing()
                     .to_owned(),
             )
-            .returning(Query::returning().columns([Embeddings::Id, Embeddings::SeqId]))
+            .returning_col(Embeddings::Id)
             .to_owned())
     }
 
@@ -79,10 +79,11 @@ impl SqliteMetadataWriter {
     ) -> Result<Option<u32>, SqliteMetadataError> {
         let (add_rec_stmt, values) =
             Self::add_record_stmt(segment_id, seq_id, user_id)?.build_sqlx(SqliteQueryBuilder);
-        let result = sqlx::query_with(&add_rec_stmt, values)
-            .fetch_one(&mut **tx)
-            .await?;
-        Ok((result.try_get::<u64, _>(1)? == seq_id).then_some(result.try_get(0)?))
+        Ok(sqlx::query_with(&add_rec_stmt, values)
+            .fetch_optional(&mut **tx)
+            .await?
+            .map(|row| row.try_get(0))
+            .transpose()?)
     }
 
     fn update_record_stmt(
@@ -107,13 +108,14 @@ impl SqliteMetadataWriter {
         segment_id: SegmentUuid,
         seq_id: u64,
         user_id: String,
-    ) -> Result<u32, SqliteMetadataError> {
+    ) -> Result<Option<u32>, SqliteMetadataError> {
         let (update_rec_stmt, values) =
             Self::update_record_stmt(segment_id, seq_id, user_id).build_sqlx(SqliteQueryBuilder);
         Ok(sqlx::query_with(&update_rec_stmt, values)
-            .fetch_one(&mut **tx)
+            .fetch_optional(&mut **tx)
             .await?
-            .try_get(0)?)
+            .map(|row| row.try_get(0))
+            .transpose()?)
     }
 
     fn upsert_record_stmt(
@@ -304,12 +306,16 @@ impl SqliteMetadataWriter {
                 Err(_) => deleted_keys.push(key),
             }
         }
-        Self::delete_metadata_by_key(tx, id, deleted_keys).await?;
-        Self::upsert_metadata(tx, id, metadata_not_null).await?;
+        if !deleted_keys.is_empty() {
+            Self::delete_metadata_by_key(tx, id, deleted_keys).await?;
+        }
+        if !metadata_not_null.is_empty() {
+            Self::upsert_metadata(tx, id, metadata_not_null).await?;
+        }
         Ok(())
     }
 
-    fn upsert_document_stmt(
+    fn add_document_stmt(
         id: u32,
         document: String,
     ) -> Result<InsertStatement, SqliteMetadataError> {
@@ -320,22 +326,17 @@ impl SqliteMetadataWriter {
                 EmbeddingFulltextSearch::StringValue,
             ])
             .values([id.into(), document.into()])?
-            .on_conflict(
-                OnConflict::column(EmbeddingFulltextSearch::Rowid)
-                    .update_column(EmbeddingFulltextSearch::StringValue)
-                    .to_owned(),
-            )
             .to_owned())
     }
 
-    async fn upsert_document(
+    async fn add_document(
         tx: &mut Transaction<'static, Sqlite>,
         id: u32,
         document: String,
     ) -> Result<(), SqliteMetadataError> {
-        let (upsert_doc_stmt, values) =
-            Self::upsert_document_stmt(id, document)?.build_sqlx(SqliteQueryBuilder);
-        sqlx::query_with(&upsert_doc_stmt, values)
+        let (add_doc_stmt, values) =
+            Self::add_document_stmt(id, document)?.build_sqlx(SqliteQueryBuilder);
+        sqlx::query_with(&add_doc_stmt, values)
             .execute(&mut **tx)
             .await?;
         Ok(())
@@ -423,20 +424,22 @@ impl SqliteMetadataWriter {
                         }
 
                         if let Some(doc) = document {
-                            Self::upsert_document(&mut tx, offset_id, doc).await?;
+                            Self::add_document(&mut tx, offset_id, doc).await?;
                         }
                     }
                 }
                 Operation::Update => {
-                    let offset_id =
-                        Self::update_record(&mut tx, segment_id, log_offset_unsigned, id).await?;
+                    if let Some(offset_id) =
+                        Self::update_record(&mut tx, segment_id, log_offset_unsigned, id).await?
+                    {
+                        if let Some(meta) = metadata {
+                            Self::update_metadata(&mut tx, offset_id, meta).await?;
+                        }
 
-                    if let Some(meta) = metadata {
-                        Self::update_metadata(&mut tx, offset_id, meta).await?;
-                    }
-
-                    if let Some(doc) = document {
-                        Self::upsert_document(&mut tx, offset_id, doc).await?;
+                        if let Some(doc) = document {
+                            Self::delete_document(&mut tx, offset_id).await?;
+                            Self::add_document(&mut tx, offset_id, doc).await?;
+                        }
                     }
                 }
                 Operation::Upsert => {
@@ -448,7 +451,8 @@ impl SqliteMetadataWriter {
                     }
 
                     if let Some(doc) = document {
-                        Self::upsert_document(&mut tx, offset_id, doc).await?;
+                        Self::delete_document(&mut tx, offset_id).await?;
+                        Self::add_document(&mut tx, offset_id, doc).await?;
                     }
                 }
                 Operation::Delete => {
@@ -750,29 +754,46 @@ mod tests {
     use chroma_types::{
         operator::{Filter, Limit, Projection, Scan},
         plan::{Count, Get},
-        BooleanOperator, CompositeExpression, DocumentExpression, MetadataComparison,
-        MetadataExpression, MetadataValue, PrimitiveOperator, Where,
+        BooleanOperator, CollectionAndSegments, CompositeExpression, DocumentExpression, LogRecord,
+        MetadataComparison, MetadataExpression, MetadataValue, OperationRecord, PrimitiveOperator,
+        Where,
     };
+    use proptest::prelude::*;
+    use tokio::runtime::Runtime;
 
-    use crate::test::TestDistributedSegment;
+    use crate::test::{TestDistributedSegment, TestReferenceSegment};
 
-    use super::SqliteMetadataReader;
+    use super::{SqliteMetadataReader, SqliteMetadataWriter};
 
-    #[tokio::test]
-    async fn test_count() {
-        let metadata_reader = SqliteMetadataReader {
-            db: get_new_sqlite_db().await,
-        };
+    proptest! {
+        #[test]
+        fn test_count(
+            operations in proptest::collection::vec(any::<OperationRecord>(), 0..100)
+        ) {
+            let runtime = Runtime::new().expect("Should be able to start tokio runtime");
+            let mut ref_seg = TestReferenceSegment::default();
+            let sqlite_seg_writer = SqliteMetadataWriter {
+                db: runtime.block_on(get_new_sqlite_db())
+            };
 
-        metadata_reader
-            .count(Count {
-                scan: Scan {
-                    collection_and_segments: TestDistributedSegment::default().into(),
-                },
-            })
-            .await
-            .expect("Count should not fail");
+            let collection_and_segments = CollectionAndSegments::test(6);
+            let seg_id = collection_and_segments.metadata_segment.id;
+
+            let logs = operations.into_iter().enumerate().map(|(i, record)| LogRecord { log_offset: (i + 1) as i64, record }).collect::<Vec<_>>();
+
+            ref_seg.apply_logs(logs.clone(), seg_id);
+            runtime.block_on(sqlite_seg_writer.apply_logs(logs, seg_id)).expect("Should be able to apply logs");
+
+            let sqlite_seg_reader = SqliteMetadataReader {
+                db: sqlite_seg_writer.db
+            };
+            let plan = Count { scan: Scan { collection_and_segments }};
+            let ref_count = ref_seg.count(plan.clone()).expect("Count should not fail");
+            let sqlite_count = runtime.block_on(sqlite_seg_reader.count(plan)).expect("Count should not fail");
+            assert_eq!(ref_count, sqlite_count);
+        }
     }
+
     #[tokio::test]
     async fn test_get() {
         let metadata_reader = SqliteMetadataReader {
